app/requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
uuid>=1.30.0
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.16
langchain-google-genai>=0.0.5
langchain-core>=0.1.5
langchain-openai>=0.1.5
pyyaml>=6.0.1
openai>=1.1.0
langgraph>=0.0.15
httpx>=0.26.0
docker>=7.0.0
